Redo routing to use hash history:
gh-pages do not support client-side browser router. For routing to work,
you have to use other routers. The problem is this application has some
legacy urls that need to stay up. This commit rewrites the router to use
hash history, with a 404 handler and a trick to rewrite old urls so that
we stay backwards compatible on gh-pages.